reduced dynamic GUI boilerplate
