+++ src/main/java/playerquests/builder/gui/dynamic/GUIDynamic.java
@@ -1,5 +1,6 @@
 package playerquests.builder.gui.dynamic;
 
+import playerquests.builder.gui.GUIBuilder; // creating the Dynamic GUI on the screen
 import playerquests.client.ClientDirector; // enables the dynamic GUI to retrieve info
 
 /**
@@ -23,6 +24,16 @@ public abstract class GUIDynamic {
     protected String previousScreen;
 
     /**
+     * If the setup has been ran
+     */
+    protected Boolean wasSetUp = false;
+
+    /**
+     * the GUI instance
+     */
+    protected GUIBuilder gui;
+
+    /**
      * Not intended to be created directly, is abstract class for dynamic GUI screens.
      * <p>
      * See docs/README for list of dynamic GUI screens.
@@ -35,7 +46,46 @@ public abstract class GUIDynamic {
     }
 
     /**
+     * Setting important values for the GUI.
+     */
+    public void setUp() {
+        this.wasSetUp = true;
+
+        // close the previous GUI
+        this.director.getGUI().getResult().close();
+
+        // create the new GUI to show the quests in
+        this.gui = new GUIBuilder(this.director);
+
+        // to-be implemented set up processes
+        setUp_custom();
+
+        // send back to the execute() body to continue
+        this.execute();
+    }
+
+    /**
      * Method to be overridden by each meta action class.
      */
-    public abstract void execute();
+    public void execute() {
+        // run setup on first time
+        if (!this.wasSetUp) {
+            this.setUp();
+            return;
+        }
+
+        // to-be implemented execute processes
+        execute_custom();
+
+        // show the results
+        if (!this.gui.getResult().isOpen()) {
+            this.gui.getResult().open(); // open GUI for first time
+        } else {
+            this.gui.getResult().draw(); // draw on the already open GUI
+        }
+    }
+
+    protected abstract void setUp_custom();
+
+    protected abstract void execute_custom();
 }
